Move files and correct pep8 comments
