+++ setup.cfg
@@ -143,4 +143,4 @@ count = False
 ignore =  # E226, E302, E41
 max-line-length = 79
 statistics = True
-exclude = migrations,static,media
+exclude = migrations,static,media,docs
