--- src/davinci_crawling/producer_consumer/__init__.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*
-# Copyright (c) 2018-2019 BuildGroup Data Services Inc.
-# All rights reserved.
-# This software is proprietary and confidential and may not under
-# any circumstances be used, copied, or distributed.
-
-"""
-Responsible for the parallelism on the crawl command.
-
-On this module is defined:
- - A multiprocess queue that will store the params generated by the
- crawl_params method.
- - A CrawConsumer that will start N threads to consume the multiprocess_queue
- and send them to the crawl method.
- - A QueueCrawler that is a modified Crawler with the add_param method
- implemented that add the parameter to the multiprocess queue.
-"""
-from multiprocessing import Queue
-
-# stores the multiprocess queue used between crawl_params and crawl method
-multiprocess_queue = Queue()
